improve code complexity of the seed
